refactor: split type helpers into feature modules
