+++ exports/main.ts
@@ -1,8 +1,13 @@
-export { createOpenApiHttp } from "../src/openapi-http.js";
-export type { HttpOptions, OpenApiHttpHandlers } from "../src/openapi-http.js";
+export type { AnyApiSpec, HttpMethod } from "../src/api-spec.js";
+
+export {
+  createOpenApiHttp,
+  type HttpHandlerFactory,
+  type HttpOptions,
+  type OpenApiHttpHandlers,
+} from "../src/openapi-http.js";
 
 export type {
-  AnyApiSpec,
-  HttpHandlerFactory,
-  HttpMethod,
-} from "../src/type-helpers.js";
+  ResponseResolver,
+  ResponseResolverInfo,
+} from "../src/response-resolver.js";
